+++ components/homepage/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Homepage Components
+ *
+ * Barrel exports for the homepage dashboard components.
+ */
+
+export * from './types'
+export { WelcomeSection } from './WelcomeSection'
+export { QuickActionsRow } from './QuickActionsRow'
+export { RecentChatsSection } from './RecentChatsSection'
+export { RecentChatItem } from './RecentChatItem'
+export { ProjectsSection } from './ProjectsSection'
+export { ProjectItem } from './ProjectItem'
+export { CharactersSection } from './CharactersSection'
+export { CharacterCard } from './CharacterCard'